Update index.html
Version 3.0: 

Proper Syntax for the most part in areas that needed clean up. There is an inclusion of an artwork of my characters that was commissioned by the artist who goes by the handle laizy_boy4. 

The final version should be version 4.0, where there will be the inclusion of a few more art/IMG assets provided by me. 

Inclusion of a video