Update howto.html
Version 3.0

Proper Syntax changes have been made for the most part. I cleaned up the code of the HTML. Most importantly, I gave webpages proper, and unique titles in the head section.

Changes to expect in the next version: 

Inclusion of a "How-To" video directly under the terms table. Video will be recorded, drawn, and voiced over by me. Expected update will be the day buy the end of day.