Update download.html
Version 3.0

This page is near complete. All assets are included in the page. All links work. Description has been written. Syntax has been checked and edited to the best version possible to where all assets work and the aesthetic is close to ideal.

Only inclusions left are the main background image behind the wrapper. Expected to be completed by the end of day.