Update FAQ.html
Version 3.0

The forum, at least at a full desktop webpage view, is close to ideal. There is room for improvement, but I would need more time. However, I have a deadline that requires I work to the best not to perfection. 

Expected edits by end of day: 
in-page links to each section of the the questions and forum to the right, and one image.