Update bookdebt.css
Version 3.0

There are some clean-ups needed for the media-quarries. However, that will have to be a personal project another time. In the meantime, I'll focus on a full screen format for my website to meet the deadline that I have set for myself today. 

Important note: Keep all images as id divs !